Okay cleared block, see note
FIrst, Added param to acceptToken as a check if a new symbol should be added.
Second, fixed an issue with the hash table array getting a copy of the objects instead of the pointer that I'm editing in the parser. Now it's an array of pointers instead of an array f objects.
Third, changed test files from book to not reference their input/output functions.